shot-captioning/requirements.txt
flask==3.0.0
torch>=2.6.0
torchvision>=0.21.0
transformers>=4.46.0
accelerate==0.27.0
pillow==10.1.0
numpy<2.0.0
pytest==7.4.3